Swaps react out for preact
