+++ components/results.js
@@ -1,3 +1,8 @@
+import { h } from 'https://cdn.pika.dev/preact@10.3.3'
+import htm from 'https://cdn.pika.dev/htm@3.0.3'
+import css from 'https://cdn.pika.dev/csz@1.2.0'
+
+const html = htm.bind(h)
 const round = num => parseFloat(num).toFixed(2)
 
 const style = {
