+++ index.html
@@ -13,6 +13,7 @@
       content="width=device-width, initial-scale=1, shrink-to-fit=no"
     />
     <script
+      defer
       async
       src="https://www.googletagmanager.com/gtag/js?id=UA-125574008-2"
     ></script>
@@ -34,7 +35,6 @@
     </script>
     <link rel="stylesheet" href="/index.css" />
     <link rel="stylesheet" href="/library/prism.css" />
-    <link rel="stylesheet" href="/routes/home/index.css" />
     <script defer src="/library/prism.js"></script>
     <script
       type="module"
