--- routes/notfound/index.js
@@ -1,6 +1,0 @@
-export default () => html`
-  <main className="missing">
-    <h1>;(</h1>
-    <a href="/">Hard Reset</a>
-  </main>
-`
